serving_pipeline/requirements.txt
streamlit
mlflow
requests
pandas
ultralytics
pillow
opencv-python
psycopg2-binary
gdown
dotenv-python 